pgwire: check for closed clients in send_rows
It may take arbitrarily long for Materialize to return rows for a query.
Possibly forever, if the query is a `SUBSCRIBE` or unchanging relation.
This commit teaches Materialize to periodically check whether the client
has hung up and, if so, terminate the session, which prevents a small
resource leak when a connection abandons a stalled query.

We do this check in the many other places where we wait for rows in the
pgwire protocol handler; we just happened to miss this place.