adapter: add internal retries (including metrics) to postgres ts oracle
This copies the persist `retry` module, because using that compared to
the one from mz_ore makes integration with metrics (which also
originally come from persist) much easier. If anyone were so inclined,
it should be easy to pull _that_ retries implementation out into a util
crate, but I didn't want to deal with yet another refactor for now.

With this change we have a bunch of duplicate metrics, for example the
"started" metrics. The "failed" and "retries" matrics also show the same
information. I think the only value that the operation-specific metrics
now bring are the "seconds" metrics, which show how long the operations
take. We could unify those metrics later, if we want.