SQL: Fix usages of SET cluster TO default
The syntax `SET var TO default` or `SET var = default`, causes var to
reset to its default value. This can cause confusion with clusters
because by default every environment has a cluster named "default". It
was common to type `SET cluster TO default` as a way to change the
cluster to the cluster named "default". However, that doesn't work in
the general case if the default value for `cluster` has been changed
for a role. It used to work fine because `"default"` was the default
value for clusters, and we hadn't implemented changing the default
value of session variables. Recently we implemented the ability to
change the default value of session variables for a role.

This commit changes all internal usages of `SET cluster TO default` to
`SET cluster TO "default"` and `SET cluster = default` to
`SET cluster = "default"`.

Works towards resolving #22351