storage: make decode_and_mfp yield configurable
We've seen evidence in production of this operator often going more than
5s without yielding, especially during rehydration and large unindexed
SELECTs. This affects interactivity as well as the time between when a
timely-driven Future is woken and when it is scheduled again.

Attempt to fix this with two tweaks to our yield heuristics. First, make
the fuel between yields configurable through LD. Second, also count each
mfp evaluation against our fuel, in case it contains a restrictive
filter.