catalog: Remove Option from epoch API
This commit updates the `epoch()` method in `DurableCatalogState` to
remove the `Option`. Opened catalogs will always have an epoch, and
will never return `None` from `epoch()`. The only way to get an
instance of `DurableCatalogState` is by opening the catalog. Therefore,
there's no reason to return an `Option` from `epoch()`.