storage-client: allow generic introspection updates
This commit adds an API to the storage controller,
`record_introspection_updates` that allows callers to record their own
updates into storage-managed collections. This is in preparation for
making the compute controller expose data in storage-managed collections
rather than builtin tables.

This commit also adapts the statement logging to use the new generic
API, rather than the specialized one it was using previously.