compute-client: check for ready messages in channels
This commit makes the compute controller's `ready` method also consider
pending messages in the two channels it shares with the per-instance
controllers. Not doing so before was an oversight.

Note that this requires us to replace the `std::sync::mpsc` channels
with crossbeam channels because the former don't support checking for
pending messages without consuming.