remove mz_words.txt (#22413)
This was intended to be a custom dictionary of words that appear in the
Materialize codebase but not in standard English dictionaries. It was
added to support spellchecking.

However, in the meantime, it has not been kept up to date, so I am
skeptical of its current usefulness. If we're not going to put in
sustained effort to keep it up to date, we might as well just delete it.