test: Deflake test_pg_cancel_dropped_role (#22410)
The test_pg_cancel_dropped_role test would start a new connection,
called dropped_client, and then use another connection, called
query_client, to look up dropped_client's session ID in mz_sessions. In
CI the lookup would occasionally fail and cause a panic. The reason is
that starting a new session and writing the session details to
mz_sessions is not linearizable, so we are not guaranteed to see it
after a session has been established. However, it does exhibit
sequential consistency, so a session will always be able to see its own
information in mz_sessions.

This commit updates test_pg_cancel_dropped_role to have dropped_client
to look up the session ID instead of query_client, which guarantees that
we will always see the ID.

Fixes #22408