Replace "symbolicate" with "symbolize" everywhere (#22412)
In the iOS ecosystem, the act of assigning names to addresses in a
backtrace is called "symbolication", so I introduced that word to
Materialize, not knowing any better. I later learned that it is instead
called "symbolization" in most other ecosystems.

This PR aligns our usage with the rest of the non-mobile-dev world.

There should be no user-visible changes (except the text changes in
prof.html, which I doubt any users are looking at anyway).