lsp-server: jinja skip + test improvements (#22252)
This PR makes the LSP extension aware of Jinja code.

Working with dbt means creating `.sql` files containing Jinja templates
that are incompatible with Materialize's parser. To avoid sending
useless error messages, the LSP server will inform the user that it is
not possible to parse Jinja content. The user can still compile the SQL
using dbt compile and open the compiled file to check for any issues.