rbac: Update view owner privilege docs (#22425)
This commit updates the docs that specify the necessary privileges of a
view owner. Materialize has the following two properties:

- Whether a role is a superuser is associated with an active database
connection/session and it is not persisted anywhere in Materialize.
- In order for someone to query a view, the view owner must have the
necessary privileges to execute the view definition.

As a consequence of these two facts, if a superuser creates a view, then
they must explicitly be granted the privileges to execute the view
definition. When another user tries to query the view, there's no way to
know that the view owner is a superuser and shouldn't need explicit
privileges.