Evolve and implement fixed-length traits for unit
This commit evolves traits related to fixed-length representations so
as to effectively accommodate specialization to empty values. Among
others, datum-based iteration is included in IntoRowByTypes to allow
for fixed-length containers, unit, and Row to use the same mechanism
uniformly at call sites. Additionally, the commit includes an
implementation of the traits for the unit type.