Introduce RowUnit arrangement specializations
This commit introduces arrangement specializations where the key is of
type Row, but the value is of type unit. The specialization process is
fairly mechanical, but of note, a specialization of Bytes9Unit is not
included. The latter is because a subsequent commit will eliminate the
9-byte specialization in order to create an early end-to-end
implementation where enable_specialized_arrangements can be safely
turned on.