Empty value specialization in Distinct reductions
This commit specializes arrangements coming out of reductions to take
advantage of empty values. In particular, the only kinds of reductions
where an empty value is produced occurs in computing Distinct. So the
approach is to wrap this output arrangement with a RowUnit tag, while
all the remaining ones employ a RowRow tag.