Remove code specific to 9-byte specialization
This commit removes all code that is specific to the 9-byte arrangement
specialization for keys. By doing so, the commit brings the remainder
of the code to a safe spot wrt. setting the
enable_specialized_arrangements flag. This is done as we'd like to get
validation for the empty value specialization prior to investing more
into key specializations. We note that trait bounds generalizing
methods in both keys and values are kept in the rendering code so that:
(a) It remains easy to bring back the removed code for key
specialization; (b) We evolve the code with the perspective that key
specialization could be attempted at short notice.