Tag arrangement names for empty values
This commit improves the names of arrangements that are created with
empty values. This can be because it is fundamental to the computation
(e.g., an intermediate step requiring distinctness) or because we have
an output arrangement created from an input arrangement with empty
value specialization. The latter happens for threshold and in
re-exports of iterative contexts; the former, in reductions.