catalog: Add shadow catalog implementation
This commit adds a shadow catalog implementation, which wraps both a
stash and a persist implementation. Each method call will dispatch the
call to both internal implementations and compare the results.

This commit also updates the majority of Rust tests and SLTs to use the
shadow implementation. Testdrive tests will be updated in a future
commit to use the shadow implementation.

Resolves #22481