proc,prof: provide extra mapping info in pprof dumps
This commit extends the pprof memory profiles we produce by a set of
`executableInfo` comments. The information in these comments is required
for offline symbolization to work correctly.

To retrieve the `executableInfo` information, the `mz-proc` crate is
extended to not just retrieve build IDs but also memory offsets and
paths of all objects loaded into the process space. This information is
sufficient to also generate the pprof `mappings`, which means that we
don't have to parse those from the jemalloc output anymore.

Note that we are now including fewer mappings in the pprof dump that we
did before this change: Only mappings of files are included, anonymous
mappings are skipped. That is fine, since for anonymous mappings we
don't have symbol information anyway, so including them into the pprof
output wouldn't give us anything.