sql: Implement `ALTER SCHEMA ... RENAME ...` (#22365)
This PR implements the ability to rename a schema.

Not all of our `create_sql` statements are normalized with an id, e.g.
`[u5 as materialize.public.foo]`, and neither are column names, e.g.
`SELECT public.foo.a FROM foo`. Because of this, to rename a schema we
need update all of the items in a schema, as well as everything that
uses these items. This PR takes exactly that approach, the core logic
uses a visitor to walk through a statement and renamed all references of
the old schema.

This feature is behind a flag, `enable_rename_schema` largely because I
still need to add audit log events. This diff was getting large enough
as-is, so I decided to do this in a follow up.

### Motivation

Progress towards
https://github.com/MaterializeInc/materialize/issues/3680

### Tips for reviewer

I broke this PR into several different commits which can be reviewed
independently:

1. Updates to the SQL parser
2. Plumbing the new statement through planning and a feature gate
3. **Logic for renaming a schema**, this is the most important commit ⭐️
4. sqllogictests
5. Parallel workload and Platform upgrade tests from QA

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered.
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [x] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
- Adds the ability to do `ALTER SCHEMA ... RENAME TO ...`, behind a
LaunchDarkly feature flag.