catalog: Implement clean shutdown for catalog state
This commit implements a method to cleanly shut down the durable
catalog state and release any of its resources. This is an async
method, otherwise it would have been added to the drop implementation.

The shutdown method is called on a best effort basis. The main usage of
the catalog is as a field in the Coordinator, which isn't ever dropped,
and therefore never cleaned up. This is OK in terms of correctness,
because anything that needs to be cleaned up is lease based, and will
just get cleaned up when the next Coordinator comes around.

However, there are some uses of the catalog, such as the stash-debug
tool, that opens the catalog and then closes it. We'd like to be able
to cleanly release all resources in that scenario without having to
wait for leases to expire.

Resolves #22467