[persist] Restore Blob state based on the contents of Consensus (#20482)
Persist state is a mix of stuff in CRDB and stuff in S3. This
complicates restoring from backup: if we restore CRDB to an old state,
the blobs it references might be deleted.

It's possible to "undelete" deleted blobs in S3 by deleting a delete
marker from an unversioned bucket. This PR implements that logic as a
CLI tool, and implements a basic test that restores a CRDB state and
checks that the database is in the expected state after the CLI runs.

### Motivation

See #18157 for the design.

### Tips for reviewer

One subtlety here is that we can't rely on the usual state iterators,
since those rely on the existence of rollups, which may not exist yet.
The current draft just iterates over the diffs manually.

I have not yet tested this tool against a real environment. I will
probably wait for at least a first pass of code review first!

Co-authored-by: Philip Stoev <pstoev@materialize.com>