catalog: Update persist catalog shard IDs (#22507)
This commit updates how the persist catalog generates shard IDs.
Previously, the persist catalog used the environment ID as the shard ID
directly. This may lead to confusion on dashboards and prevents the
persist catalog from using multiple shards. This commit updates the
shard generation so that it's deterministically generated from the
environment ID.

Resolves #22506