expr: add `MirScalarExpr` helper functions
Add a `MirScalarExpr::is_column` helper function that can be used to
classify trivial expression sequences.

Add a `MirScalarExpr::might_error` helper function to be used as a
non-complete classifier for predicates that we might want to move around
because we don't want to change the error semantics of the `MirRelationExpr`.