compute: mz_dataflow_delayed_time_seconds_total metric
This commit adds a metric tracking a replica's "delayed time", i.e. the
time during which at least one of the dataflows installed on the replica
hasn't been fully caught up. Our hope is that this will be a useful
metric to estimate replica utilization, even in the face of idle
merging.

The metric is calculated based on dataflow import and export frontiers.
The compute logging dataflow has access to both, so it is made
responsible for maintaining the metric.

The main wrinkle with this new metric is that it is based on import
frontiers, which means it doesn't account for:

 * the time between dataflow creation and when the first import frontier
   is logged
 * logging dataflows that have no imports and therefore no import
   frontiers