persist: wrap shard_source open_leased_reader call in a tokio task
Internally, the `open_leased_reader` call registers a new LeasedReaderId
and then fires up a background tokio task to heartbeat it. It is
possible that we might get a particularly adversarial scheduling where
the CRDB query to register the id is sent and then our Future is not
polled again for a long time, resulting is us never spawning the
heartbeat task. Run reader creation in a task to attempt to defend
against this.

h/t Nikhil for the theory