test: Parallelize test_builtin_migration
test_builtin_migration is a notoriously slow test. Previously, in CI it
took roughly 120s. The main bottleneck of this test is constantly
opening a new catalog, which is a slow process. The test is testing the
process of opening the catalog, so there's not much we can do to avoid
this step.

This commit breaks up test_builtin_migration into multiple tests to
allow easier parallelization of subtests. Keeping all subtests in a
single test, but running them in different tasks does not result in
any speedup. All tests get serialized by CRDB transactions since
opening the catalog requires a CRDB transaction that will conflict
with any other catalog opening transactions.