stash: Use transaction priority to avoid aborts
The stash has two concurrent tasks executing transactions against
CockroachDB. One of the tasks is executing the logic for DDL and the
other is performing compaction in the background. Occasionally these
two tasks will execute transactions at the same time causing one to
abort. Executing logic for DDL is much more important than compaction,
compaction is not needed for correctness and is only used to reclaim
unused data and speed up reads.

This commit executes the main stash task as a high priority CockroachDB
transaction and compaction as a low priority CockroachDB transaction.
The result is that if there's a conflict, the compaction task is always
the one that's aborted. See:
https://www.cockroachlabs.com/docs/stable/architecture/transaction-layer#transaction-conflicts

Maybe works towards fixing #19931