test: Speed up test_subscribe_progress (#22633)
test_subscribe_progress iterates on a couple of different parameters and
executes roughly the same test for each variant of each parameter.
Previously, each iteration would stop and start a new Materialize
server. Considering that starting a new server is fairly expensive, this
cause test_subscribe_progress to be very slow. Roughly 94.158s in CI and
160s locally on my machine. There's no correctness reason for each
iteration to start with a fresh server.

This commit updates test_subscribe_progress so that it only starts a
single server and uses it for each iteration. Locally this caused the
test to speed up to 37s. In CI it sped up to 12.054s.