compute: use granular backpressure in persist_source decoding (#22549)
We've seen instances during rehydration in prod of OOMs from what
appears to be a backup of fetched persist parts waiting on decoding. The
recent change to drive s3 work with tokio instead of timely has likely
eliminated any tiny bit of backpressure we were getting from timely
scheduling, and this was already plausible.

In a way, decoding is quite analogous to writing upsert data to rocksdb,
so reuse the recent fine-grained backpressure work from that. I did some
testing on Friday and the results are quite promising. Specifically:

- Turned on the TPCH loadgen source
- Created all 21 TPCH queries as queries in a cluster
- Added and removed replicas of various sizes and watched rehydration

Before, we saw memory spikes during rehydration of 60-70 GiB+, settling
down to 8-9 GiB. With this change and a value of 128 MiB, we saw memory
spikes of 8-13 GiB before settling down to the same 8-9 GiB.

![image](https://github.com/MaterializeInc/materialize/assets/52528/544ed7cf-0a99-4681-a714-10a4c2d8780d)

Note that the recording interval for memory usage here is every 60s, so
the real spikes might be bigger, but the numbers are dramatic enough
that the impact is almost certainly real. I did also try an xsmall (13.5
GiB) and it consistently OOMed, confirming this point.

Ideally, this parameter would be tunable in a more targeted way
(per-cluster/replica/persist_source), but that's much harder. We already
had this LD knob, and the affected customers aren't the most sensitive
to rehydration time, so confirm that this fixes things before investing
more heavily here. This repurposes the `dataflow_max_inflight_bytes`
config from end-to-end compute dataflow backpressure to instead be
granular backpressure purely inside compute's persist_source instances
holding back fetching on decoding. This config isn't even created in LD,
so there aren't any concerns with reusing it.

- Note that this doesn't prevent us from re-introducing the end-to-end
dataflow backpressure in the future. Also note that with this PR, the
remaining benefit of end-to-end backpressure is really catching up when
a computation has been down for a while (the inputs are far ahead of the
output). But, this has a nice workaround (temporarily add a larger
replica, spin it down when things are caught up) whereas the issue this
PR is solving doesn't really have a good workaround (the replica must be
oversized just to survive rehydration).

Touches https://github.com/MaterializeInc/cloud/issues/7835

### Motivation

  * This PR fixes a recognized bug.