environmentd: silence orchestrator in integration tests
This (mostly) disables clusterd INFO messages, which are largely spam
for envd tests and have greatly decreased our ability to quickly
understand test output.

This still produces occasional WARN messages I think because the test
macro we use auto-inits the tracing subscriber, so we are not able to
use the normal TracingCliArgs stuff. That may be a refactor for another
day.