catalog: Open catalog in single transaction
Previously, the opening the catalog involved multiple independent reads
and writes from and into the durable catalog store. This commit updates
opening the catalog, so that it's all done in a single catalog
transaction. As a consequence, there's only a single atomic read and
write from and into the durable catalog store. This change has the
following benefits:

  - Opening the catalog is easier to reason about since it's a single
  atomic change to the durable catalog store, instead of many
  independent changes. This is especially true if opening the catalog
  fails part way through.
  - Opening the catalog is faster since there's less communication with
  the durable catalog. Note this is untested.
  - Many of the `DurableCatalogStore` and `ReadOnlyDurableCatalogStore`
  methods can be removed, which decreases the code that is specific to
  a single implementation.

In order to accomplish opening the catalog in a single transaction, we
had to make the following concrete changes:

  - Move some functionality from `Catalog` to `CatalogState`.
  - Move some functionality from `DurableCatalogState` and
  `ReadOnlyDurableCatalogState` to `Transaction`.
  - Update the methods responsible for opening to catalog from working
  with a `Catalog` to working with a `CatalogState` and `Transaction`.

Resolves #22667