sql: add HIR ⇒ MIR lowering config
Create a dedicated `Config` struct to host feature flags that affect
the HIR ⇒ MIR lowering step. Currently this config has only one flag
which determines the boundary between the equijoin-based and the
fallback outer join lowering strategy.

Most of the changes in this commit are threading a new `config: &Config`
parameter called through the lowering code. In addition, we also change
the definition of `OnPredicates::is_equijoin` to be conditioned on the
value of the `enable_new_outer_join_lowering` config parameter.

The parameter is still not wired through the client-facing `lower()`
call (this will happen in a follow-up commit).