[process orchestrator] Use systemd if available
This PR uses systemd-run if available to limit the CPU and memory resources
a process can consume. It detects at runtime if systemd is available, and
defaults to the current behavior if it's not.

Also adjusts the replica size definitions to include a CPU limit, and
introduce some variants with explicit memory limits. We could align this
to be closer to what we use in production.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>