adapter: add an initial read timestamp when creating a table
The current code does:

- get_local_write_ts
- register (which doesn't use the timestamp)
- group commit (which does use the timestamp)
- apply_local_write

When the new txns stuff is turned on, the register call _will_ use the
timestamp and the group commit will become unnecessary. We could guard
it with an check for whether we're using the new txns stuff, but this
would be the _only_ place outside of the storage controller that would
have to know about it. Instead, just grab a second timestamp for the
group commit. Create tables should be rare and at most one of them will
actually need to communicate with CRDB since we still allocate ranges of
timestamps. Then, the group commit can go away entirely once we've
finished switching over.