storage: impl TxnsCodec for SourceData
This means the txns shard will be storing its data encoded in the usual
`<SourceData, (), u64, i64>` format, so we should be able to hook it up
as an internal source for debugging.

Also implement the mz_persist_types::StepForward copy of `step_forward`.
There's probably something we could do here to deduplicate them, but
it's not clear yet that it would be better, so hold off for now.