*: hook up new persist-txn tables implementation behind a flag
This is primarily in service of Pv2's use-case isolation. Switching our
tables to persist-txn is necessary, but not sufficient work to tease
apart table txns into isolated units.

The persist-txn library presents an abstraction on top of persist
shards, allowing efficient atomic multi-shard writes. This is
accomplished through an additional _txns_ shard that coordinates writes
to a (potentially large) number of _data_ shards. Data shards may be
added to and removed from the set at any time.

From Pv2's perspective the motivating benefit is that these multi-shard
txns are _atomic_. This will allow us to build an Optimistic Concurrency
Control protocol for our SQL transactions, removing the single-process
global write lock.

Eventual secondary benefits of persist-txn tables will be:
- Greatly decreased Consensus/CRDB write traffic. Currently, tables
  require `O(num_tables * passage_of_time)` writes. With persist_txn,
  that will be drastically reduced to `O(num_tables_in_txn) +
  O(passage_of_time)`. That is, before we had to write to each table
  once per second, after we will write one write per second plus one for
  each table involved in a txn. This will hopefully allow us to
  essentially remove the system limit on the number of tables.
- Bounded memory table writes. Technically, this is possible today, but
  the division of responsibilities between adaptor and storage makes it
  tricky. The persist-txn API will make it much easier to confidently
  surface lower-level table details up to adaptor.

I was really hoping to get this branch to a point where I could merge it
with the new behavior on-by-default in CI, but a number of work items
have turned out be better handled as post-integration followups. These
are tracked in #22801. Notably, many of these are performance related
(crdb writes are down, crdb and s3 reads are up, a couple extra ops in
the INSERT path) and we haven't finished fleshing out the story for
flipping the feature flag off.

Touches #22173