catalog: Optimize catalog initialization (#22709)
`TableTransaction`s provide a method called `set` that allows someone to
set the value of a specific key. After setting the value, the method
runs a uniqueness check on all keys and values. Certain code paths call
`set` in a loop, causing the uniqueness check to be run multiple times.

This commit adds a new method to `TableTransaction` called `set_many`,
that sets the value of many keys and then runs a single uniqueness
check. This allows us to reduce latency in the cases where we were
calling `set` in a loop.