explain: support `enable_new_outer_join_lowering` flag
Plumb through the `ExplainConfig` through `explain_statement` in order
to handle the `enable_new_outer_join_lowering` in code that explains a
`SELECT`, `CREATE MATERIALIZED VIEW`, or `CREATE INDEX` statements.

I had to modify the global optimization pass for these by adding a block
that runs only in `EXPLAIN` mode and re-does the `local` optimization
for each imported `view` item in the `DataflowDescription` before those
are inlined. This should be good because local optimization depends only
on its configuration and not on other runtime context.