stash: Optimize TableTransaction set and set_many (#22736)
This commit optimizes the `set` and `set_many` methods in
`TableTransaction` so that no action is taken if the provided values are
equal to the existing values.

This has the added benefit that if `set` or `set_many` are called in a
read-only context, and they make no changes, then they will not cause a
read-only error. This matches the semantics of other stash methods such
as `TypedCollection::upsert_key`.