Re-use row allocation
This commit changes how we construct rows in the TPC-H load generator.
Instead of using `Row::pack_slice`, we retain an allocation that we reuse
when creating a new row. This avoids having to calculate the size of the
row as the buffer will have enough capacity after a few iterations.

This is based on me seeing this code path standing out in a perf run. The
only downside is increased code complexity and that we retain two row
allocations, which should be a marginal increase in the case of the TPC-H
load generator.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>