persist: change downcast_stats error! log to warn!
We saw this pop up, which means the error! has done its job. While we
investigate, downgrade the log to a warn! to avoid spamming sentry. Add
metrics coverage to compensate, so we don't have to rely on searching
logs to see if it's still happening.

While I'm in here, add some context to the message so it's easier to
track down next time (I had to do a lot of manual correlating during the
initial triage).

Touches #22680