sql[-parser]: use standard option parsing/planning for ASSERT NOT NULL
As discussed in [0], the recently added `ASSERT NOT NULL` option should
use the new standard option parsing/planning technique (name enums and
the `generate_extracted_config` option).

There is one minor behavioral change in this commit: the `ASSERT NOT
NULL` option can now optionally be followed by an equals sign, for
consistency with other `WITH` options, as in:

    CREATE MATERIALIZED VIEW v WITH (
        ASSERT NOT NULL = c1
    ) AS SELECT 1 AS c1

The bulk of the code changes here are teaching the
`generate_extracted_config` macro to handle options that can be
specified multiple times.

[0]: https://www.notion.so/SQL-design-principles-2e2069797f7c46b8b4c7dd10043afdcd?d=25caa2d22e514d1c80f1fab4cd4ae7aa&pvs=4#10673ea6695840e8addcf0ac73bb7392