backup: Add additional scenarios for backup+restore
Add additional backup+restore scenarios, including one where
Mz is allowed to continue to run and execute queries after the backup
has been taken.

To accomodate the scenario, which needs to run the manipulate(#2) phase
once again after the restore, the `@externally_idempotent(False)` decorator
is introduced to annotate Checks that perform actions against third-party
services that are not idempotent.