Add github action to publish sql parser changes to the sql council slack (#22798)
This PR adds a GitHub Actions workflow that pushes a notification to the
#epd-sql-council Slack channel when the SQL parser is changed in a PR.
PRs touch this file ~10-15 times/month and the volume is declining, so
I'm hoping the notifications are useful and manageable, but I can turn
this off if they're not.

Part of https://github.com/MaterializeInc/eng-org/issues/19.

### Motivation

* This PR adds new automation.

### Tips for reviewer

GitHub does not allow actions invoked through the `pull_request` event
to access secrets, like the Slack token. This workflow uses
`pull_request_target`, which unfortunately means that we have to merge
it to main before we can actually test it. Are there any ways I can
partially test before merging?