catalog: Set connect timeout during transactions (#22802)
This commit allows setting the catalog connect timeout during a
transaction. This allows you to set the connect timeout that will be
used for the transaction itself during commit. This is especially
helpful when opening the catalog, where we want to use the connect
timeout, but we need to start a transaction to fetch it.