persist: plumb down an override for the next_listen_batch sleep params
The next_listen_batch sleeps are now only a fallback for pubsub edge
cases (crashes, full buffers etc). We've also got them tuned quite
directly for the current write pattern of the upper advancing once per
second. In particular, an initial sleep of just over a second (so most
times it's preempted by a pubsub wakeup and we never get to the crdb
state scan) and then every 100ms after (if pubsub didn't wake it up
within a second, it probably won't and we want to stay responsive).

Shards managed by persist-txn break this "upper advances every second"
assumption by allowing the "logical" upper to advance past the
"physical" upper when the shard is not being written to. However, this
plays poorly with the above listen tunings: after just over a second,
they start polling every 100ms until the next time the shard is written,
which for most system tables is only on restart.

This PR plumbs an override for the listen sleep retry parameters all the
way from persist_source_core down to next_listen_batch, so that we can
override the behavior for ones paired with a txns_progress operator.